Firebase Update, Explore All Update, Filter Update, Navbar Update
Tried Adding a JS file that auto update the Firebase every change, The design is similar to an API. Every time a change in the firebase is needed we can change it in the dormitories.js. If necessary I can also add user.js but for now no.

Completed the Explore All Page with working filters and CSS. To change the description, change it inside dormitories.js and it will update it immediately

Added view buttons. Cards, and design.

Updated The Navbar where the user can go back to the main page by pressing the logo.

Note: A lot from the dormitory database is are all place holders.