Dormitory Manager Dashboard (v1.0)
This hefty change accounts for two out of the four major sections of the Dormitory Manager Dashboard, namely:

Rooms
Dormers

Most of the data fetching is dynamic, wherein data need not be stored inside a local folder.

Kindly take time to monitor and report malfunctions so that it could be resolved within the time we have to prepare for the final presentation.